justfile: Read agenix key from keychain on macOS
That's a lot better than putting it somewhere on disk.

We could probably do something like that for Linux too. It's not that
bad on Linux since the default location is likely an in-memory
filesystem but it's still more convenient to use the keyring daemon
instead of having to copy stuff to a file manually after rebooting.