justfile: Fix `generate-secrets` recipe on macOS
Prior to this commit, the admin key wasn't loaded from the keychain and
it wouldn't work at all since `XDG_RUNTIME_DIR` isn't a thing on Mac.

The code responsible for locating the keyfile is currently duplicated
in the `agenix` recipe but ehhh I don't wanna take care of that right
now.