Integrate mikubook 🥳
This is the first nix-darwin host and the first non-NixOS host to be
integrated into this repo. And globalmods just work, as expected~ :3

The nix-darwin config could be improved, the integration in the
"configurations" flake module is still super kludgy but it's there and
it works, so I'm committing it.

Side note: It's called mikubook bc it's a blue MacBook Air with a
sticker of Hatsune Miku flying the trans flag on it. <3